moved ICEsat2 SI tools and configuration files
