+++ .gitignore
@@ -39,6 +39,9 @@ analysis_db/support_files/
 .installed.cfg
 *.egg
 
+*__pycache__/
+*__pycache__/*
+
 # Environments
 .env
 .venv
