+++ src/icesat2_tracks/ICEsat2_SI_tools/__init__.py
@@ -1,0 +1,1 @@
+__version__ = "0.0.0"
