+++ src/icesat2_tracks/ICEsat2_SI_tools/angle_optimizer.py
