+++ src/icesat2_tracks/ICEsat2_SI_tools/filter_regrid.py
