+++ src/icesat2_tracks/ICEsat2_SI_tools/spectral_estimates.py
