+++ src/icesat2_tracks/ICEsat2_SI_tools/spicke_remover.py
