+++ src/icesat2_tracks/config/IceSAT2_startup.py
@@ -15,7 +15,10 @@ import matplotlib.pyplot as plt
 from matplotlib.gridspec import GridSpec
 import matplotlib.colors as colors
 import pandas as pd
-
+from . import utils as utils
+from icesat2_tracks.local_modules import m_colormanager_ph3 as M_color
+from icesat2_tracks.local_modules import m_tools_ph3 as MT
+from icesat2_tracks.local_modules import m_general_ph3 as M
 
 import string
 
@@ -25,6 +28,7 @@ import xarray as xr
 
 import sys
 import imp
+import json
 
 
 
@@ -32,25 +36,30 @@ import imp
 #import m_general as M
 
 #import AA_plot_base as AA
-def json_load(name, path, verbose=False):
-    import json
-    full_name= (os.path.join(path,name+ '.json'))
+# def json_load(name, path, verbose=False):
+   
+#     full_name= (os.path.join(path,name+ '.json'))
 
-    with open(full_name, 'r') as ifile:
-        data=json.load(ifile)
-    if verbose:
-        print('loaded from: ',full_name)
-    return data
+#     with open(full_name, 'r') as ifile:
+#         data=json.load(ifile)
+#     if verbose:
+#         print('loaded from: ',full_name)
+#     return data
 
-mconfig=json_load('config','/gpfs/data/chorvat/IS2/2021_ICESat2_tracks/config/')
+print("HELLO FROM IceSat2 startup 6666666666666666")
+
+mconfig=utils.json_load('config','/mnt/c/Projects/Github/icesat2-tracks/config/')
 
 # add project depenent libraries
-sys.path.append(mconfig['paths']['local_script'])
-sys.path.append(mconfig['paths']['local_script'] +'/ICEsat2_SI_tools/')
+# sys.path.append(mconfig['paths']['local_script'])
+# sys.path.append(mconfig['paths']['local_script'] +'/ICEsat2_SI_tools/')
 
-import m_colormanager_ph3 as M_color
-import m_tools_ph3 as MT
-import m_general_ph3 as M
+
+# import m_colormanager_ph3 as M_color
+
+print("M_color path"+M_color.__file__)
+# import m_tools_ph3 as MT
+# import m_general_ph3 as M
 
 #load colorscheme
 col=M_color.color(path=mconfig['paths']['config'], name='color_def')
@@ -239,3 +248,17 @@ def font_for_pres():
 
 # add project depenent libraries
 #sys.path.append(config['paths']['local_script'])
+
+# class icesat2_confi:
+#     _instance = None
+#     def __new__(cls):
+#         if cls._instance is None:
+#             STARTUP_2021_IceSAT2="config/2021_IceSAT2_startup.py"
+#             exec(open(STARTUP_2021_IceSAT2).read())
+#             cls._instance = super(icesat2_confi, cls).__new__(cls)
+#         return cls._instance
+
+def my_function():
+    print("HELLO FROM IceSat2 startup")
+
+my_function()
