+++ src/icesat2_tracks/local_modules/__init__.py
@@ -1,0 +1,1 @@
+__version__ = "0.0.0"
