+++ src/icesat2_tracks/local_modules/m_general_ph3.py
@@ -22,7 +22,8 @@ from matplotlib.colors import LinearSegmentedColormap
 from matplotlib import dates
 import datetime as DT
 #from cdo import *   # python version
-import m_tools_ph3 as MT
+#from Chess import <class-name>
+from . import m_tools_ph3 as MT
 import numpy as np
 import matplotlib.pyplot as plt
